expose some more headers
