+++ rmw_fastrtps_cpp/include/rmw_fastrtps_cpp/custom_publisher_info.hpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef TYPES__CUSTOM_PUBLISHER_INFO_HPP_
-#define TYPES__CUSTOM_PUBLISHER_INFO_HPP_
+#ifndef RMW_FASTRTPS_CPP__CUSTOM_PUBLISHER_INFO_HPP_
+#define RMW_FASTRTPS_CPP__CUSTOM_PUBLISHER_INFO_HPP_
 
 #include "fastrtps/publisher/Publisher.h"
 
@@ -27,4 +27,4 @@ typedef struct CustomPublisherInfo
   const char * typesupport_identifier_;
 } CustomPublisherInfo;
 
-#endif  // TYPES__CUSTOM_PUBLISHER_INFO_HPP_
+#endif  // RMW_FASTRTPS_CPP__CUSTOM_PUBLISHER_INFO_HPP_
