+++ rmw_fastrtps_cpp/src/rmw_get_gid_for_publisher.cpp
@@ -16,8 +16,8 @@
 #include "rmw/rmw.h"
 #include "rmw/types.h"
 
+#include "rmw_fastrtps_cpp/custom_publisher_info.hpp"
 #include "rmw_fastrtps_cpp/identifier.hpp"
-#include "types/custom_publisher_info.hpp"
 
 extern "C"
 {
