+++ rmw_fastrtps_cpp/src/rmw_subscription.cpp
@@ -27,7 +27,7 @@
 #include "namespace_prefix.hpp"
 #include "qos.hpp"
 #include "rmw_fastrtps_cpp/custom_participant_info.hpp"
-#include "types/custom_subscriber_info.hpp"
+#include "rmw_fastrtps_cpp/custom_subscriber_info.hpp"
 #include "type_support_common.hpp"
 
 extern "C"
